+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - **[CHANGE]** (`ktx-actors`) `onTouchEvent` parameters renamed from `downListener` and `upListener` to `onDown` and `onUp`.
 - **[CHANGE]** (`ktx-actors`) Event listeners attached with extension methods now consume the `Actor` as `this`.
 Listeners that used to consume actors as regular parameters now should rely on `this` instead.
+- **[FEATURE]** (`ktx-actors`) `KtxApplicationAdapter`, `KtxGame` and `KtxScreen` now explicitly implement the `Disposable` interface,
+matching the `dispose` method from `ApplicationListener` and `Screen` respectively. This allows to leverage existing `Disposable` utilities.
 - **[FEATURE]** (`ktx-ashley`) Added `Engine.configureEntity` extension method that allows to add components to an existing entity.
 - **[FEATURE]** (`ktx-graphics`) Added support for `Color` destructuring syntax.
 - **[CHANGE]** (`ktx-scene2d`) Deprecated top-level widget factory methods and tooltip utilities were removed.